+++ Stitch/Graph/StitchAI/Mapping/ActionsToSyntax/deriveSyntaxFromLayerInput.swift
@@ -1,70 +1,18 @@
 //
-//  ViewModifierToLayerInputMapping.swift
+//  deriveSyntaxFromLayerInput.swift
 //  Stitch
 //
-//  Created on 6/23/25.
+//  Created by Christian J Clampitt on 6/24/25.
 //
 
 import Foundation
-import StitchSchemaKit
-import SwiftUI
 
-
-// TODO: combine these functions, so that you can make sure you capture or handle ALL `LayerInputPort` cases? i.e. what is important here
-
-struct EdgeOrigin: Equatable, Codable, Hashable {
-    let id: NodeId
-    let port: Int
-}
-
-enum StitchValueOrEdge: Equatable {
-    // a manually-set, always-scalar value
-    case value(PortValue)
-    
-    // an incoming edge
-    case edge(NodeId, Int) // `from node id + from port`
-    
-    var asSwiftUILiteralOrVariable: String {
-        switch self {
-        case .value(let x):
-            return x.asSwiftUILiteral
-        case .edge(let x, let y):
-            // TODO: JUNE 24: probably will use edge.origin's NodeId and Int to ... what? look up a proper variable name? ... Can see how Elliot does it?
-            return "x"
-        }
-    }
-    
-    var asSwiftSyntaxKind: SyntaxArgumentKind {
-        switch self {
-        case .value:
-            // TODO: JUNE 24: Do you need all these different individual syntax-literal types? ... if so, then should map on
-            return SyntaxArgumentKind.literal(.string)
-        case .edge(let x, let y):
-            // TODO: JUNE 24: do we always want to treat an incoming edge as a variable ? ... a variable is just an evaluated expression ?
-            return SyntaxArgumentKind.variable(.identifier) // `x`
-        }
-    }
-}
-
-extension PortValue {
-    // TODO: JUNE 24: should return a Codable ? ... How do you map between Swift/SwiftUI types and Stitch's PortValue types ? ... see note in `LayerInputPort.toSwiftUI`
-    var asSwiftUILiteral: String {
-        switch self {
-        case .number(let x):
-            return x.description
-        case .string(let x):
-            return x.string
-        default:
-            return "IMPLEMENT ME: \(self.display)"
-        }
-    }
-}
 
 // TODO: JUNE 24: how to handle loops? ... note: this is not used anywhere yet
 extension LayerInputPort {
-    func toSwiftUI(_ viewNode: SyntaxView,
+    func toSwiftUISyntax(_ viewNode: SyntaxView,
                    port: StitchValueOrEdge, // loops? should pass in `value` ?
-                   layer: Layer) -> FromLayerInputToSwiftUI {
+                   layer: Layer) -> FromLayerInputToSyntax {
         
         // TODO: JUNE 24: ASSUMES SINGLE-PARAMETER PORT VALUE, i.e. .opacity but not .size
         let buildModifier = { (kind: SyntaxViewModifierName) -> SyntaxViewModifier in
@@ -307,185 +255,5 @@ extension LayerInputPort {
             
             return .unsupported
         }
-    }
-}
-
-extension SyntaxViewConstructorArgument {
-    func toLayerInput(_ layer: Layer) -> LayerInputPort? {
-        switch self.label {
-            
-        case .systemName:
-            return .sfSymbol
-            
-        // TODO: JUNE 24: *many* SwiftUU
-        case .unlabeled:
-            switch layer {
-            case .text, .textField:
-                return .text
-            default:
-                return nil
-            }
-            
-            
-        case .unsupported:
-            return nil
-        }
-    }
-}
-
-extension SyntaxViewModifierName {
-    func toLayerInput(_ layer: Layer) -> LayerInputPort? {
-        switch (self, layer) {
-            // Universal modifiers (same for every layer)
-        case (.scaleEffect, _):
-            return .scale
-        case (.opacity, _):
-            return .opacity
-        case (.offset, _):
-            return .position
-        case (.rotationEffect, _):
-            return .rotationZ
-        case (.rotation3DEffect, _):
-            // Depending on the axis specified in the arguments
-            // This would need argument extraction to determine X, Y, or Z
-            return .rotationZ // Default to Z rotation
-        case (.blur, _):
-            return .blurRadius
-        case (.blendMode, _):
-            return .blendMode
-        case (.brightness, _):
-            return .brightness
-        case (.colorInvert, _):
-            return .colorInvert
-        case (.contrast, _):
-            return .contrast
-        case (.hueRotation, _):
-            return .hueRotation
-        case (.saturation, _):
-            return .saturation
-        
-        case (.fill, _): // fill is always color
-            return .color
-            
-            //    case (.font, .text):
-            //        return .font
-            
-            //    case (.fontWeight, _):
-            //        //            return .fontWeight
-            //        return nil
-            
-            //    case (.lineSpacing, _):
-            //        return nil // return .lineSpacing
-            
-        case (.cornerRadius, _):
-            return .cornerRadius
-            
-            //    case (.shadow, _):
-            //        // Shadow would need to be broken down into multiple inputs:
-            //        // .shadowColor, .shadowRadius, .shadowOffset, .shadowOpacity
-            //        return .shadowRadius
-            
-        case (.position, _): return .position
-            
-        // TODO: JUNE 23: .frame modifier is actually several different LayerInputPort cases: .size, .minSize, .maxSize
-        case (.frame, _):
-            return .size
-            
-        case (.padding, _):
-            return .padding // vs .layerPadding ?!
-            
-        case (.zIndex, _):
-            return .zIndex
-        
-        case (.foregroundColor, let kind) where kind != .text && kind != .textField:
-            return .color
-
-        case (.foregroundColor, _):
-            return .color
-
-            //        case (.backgroundColor, _):
-            //            return .color
-                        
-        case (.disabled, _): return nil
-        case (.background, _): return nil
-        case (.font, _): return nil
-        case (.multilineTextAlignment, _): return nil
-        case (.underline, _): return nil
-        case (.keyboardType, _): return .keyboardType
-        case (.disableAutocorrection, _): return nil
-        case (.clipped, _): return .clipped // return .isClipped
-        case (.custom(_), _): return nil
-        }
-    }
-}
-
-/// Maps ModifierKind cases to LayerInputPort cases based on the Layer type
-/// Returns the appropriate LayerInputPort for a given ModifierKind and Layer combination
-/// - Parameters:
-///   - modifier: The ModifierKind to map
-///   - layer: The Layer type to map for
-/// - Returns: The corresponding LayerInputPort if a mapping exists, nil otherwise
-func mapModifierToLayerInput(modifier: SyntaxViewModifierName,
-                             layer: Layer) -> LayerInputPort? {
-    
-    switch (modifier, layer) {
-        // Universal modifiers (same for every layer)
-    case (.scaleEffect, _):
-        return .scale
-    case (.opacity, _):
-        return .opacity
-    case (.offset, _):
-        return .position
-    case (.rotationEffect, _):
-        return .rotationZ
-    case (.rotation3DEffect, _):
-        // Depending on the axis specified in the arguments
-        // This would need argument extraction to determine X, Y, or Z
-        return .rotationZ // Default to Z rotation
-    case (.blur, _):
-        return .blurRadius
-    case (.blendMode, _):
-        return .blendMode
-    case (.brightness, _):
-        return .brightness
-    case (.colorInvert, _):
-        return .colorInvert
-    case (.contrast, _):
-        return .contrast
-    case (.hueRotation, _):
-        return .hueRotation
-    case (.saturation, _):
-        return .saturation
-        
-        // Layer-specific modifiers
-    case (.foregroundColor, .text):
-        return .color
-        
-    case (.foregroundColor, let kind) where kind != .text:
-        return .color
-        
-    case (.fill, _): // fill is always color
-        return .color
-        
-        //    case (.font, .text):
-        //        return .font
-        
-        //    case (.fontWeight, _):
-        //        //            return .fontWeight
-        //        return nil
-        
-        //    case (.lineSpacing, _):
-        //        return nil // return .lineSpacing
-        
-    case (.cornerRadius, _):
-        return .cornerRadius
-        
-        //    case (.shadow, _):
-        //        // Shadow would need to be broken down into multiple inputs:
-        //        // .shadowColor, .shadowRadius, .shadowOffset, .shadowOpacity
-        //        return .shadowRadius
-        
-    default:
-        return nil
     }
 }
